IDE-834 should be using getInt(.) in doLoad() vs getDouble(.)

+++ plugins/gov.redhawk.ui.port.nxmplot/src/gov/redhawk/internal/ui/preferences/OverridableIntegerFieldEditor.java
@@ -132,7 +132,7 @@ public class OverridableIntegerFieldEditor extends IntegerFieldEditor {
 		if (textControl != null) {
 			String str;
 			if (override) {
-				str = "" + getPreferenceStore().getDouble(getPreferenceName());
+				str = "" + getPreferenceStore().getInt(getPreferenceName());
 			} else {
 				str = getAutoValueForText();
 			}
